ADDed capacitive touch pins capabilities
Eight GPIOs with capacitive touch capabilities provide raw data to the
logic. These raw analog signal data values are conditioned by the logic
to provide true/false values to represent the contact with the skin for
each of the pins. This is achieved by calculating average baseline
values for each of the inputs during the power-up of the device. This
set of baseline values allow for a proper detection of the actual
touch/no touch condition of the circuits.

In the hardware, circuits for each of these pins are arranged to
provide an area of horizontal naked wires that simulate the "touchpad"
that will eventually be implemented in the ring controller.