ADDed bluetooth config and functions
A new testing Arduino project was added with everything needed
to configure and send data via bluetooth. A set of functions to
simulate the sending of a pattern of gestures were added, as well
as the function to send individual gestures symbols that will be
used when the gesture interpreter is ready to be used.

With these additions now it is possible to test the frontend end
of the bluetooth communication and simulate a user interaction
with the graphical interface.