ADDed WiFi connection config functions
As an alternative to the bluetooth connection, another testing
Arduino project was added this time with the first steps for
configuring a TCP/IP connection over WiFi.

So far the prototype would be able to connect to a wifi network,
obtain a dynamic IP address and report it to the serial debugging
console.

The next step would be to open an UDP socket and send characters
over it.